extract integration tests to a separate package
